+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Add and improve tests for API ([#168])
     - Add test cases covering the content of GeoJSON properties.
     - Improve API response schemata by using less logic to create the schemata.
+- Load indicator from DB will also load its data attributes ([#179])
 
 ### How to upgrade?
 
@@ -30,6 +31,7 @@
 [#149]: https://github.com/GIScience/ohsome-quality-analyst/pull/149
 [#153]: https://github.com/GIScience/ohsome-quality-analyst/pull/153
 [#168]: https://github.com/GIScience/ohsome-quality-analyst/pull/168
+[#179]: https://github.com/GIScience/ohsome-quality-analyst/pull/179
 [`pydantic`]: https://pydantic-docs.helpmanual.io/